feat(*): ZSH configuration improvement
Utils moved from default.nix to modules/shell/zsh.nix, for a better
modularity. .zshrc is part of the XDG file, next step is to fetch from
git repository of dotfiles (not planning to give up on that).

GNU Privacy Guard and Pass configurations using the default directories
instead of the XDG specification (at least for now, to keep backwards
compatibility with Arch Linux config).