feat(xdg): Extensive support for XDG spec
XDG (almost) all the things!

Mostly modules with customizations (through config files) are
migrated to support XDG. For the ones that aren't yet: your time is
about to come.

Additionally a huge refactor in the ZSH module took place to include
support for Antigen and Oh-my-zsh together. Stored inside XDG config
home to be complaint