chore(*): Usual Nix Flake update
Ran the usual nix flake update to gather the latest packages; updates
the lockfile;

Added usbutils to Android module and enable it back on NixOS cosmos
(nixosmos);

Once PyYAML was updated at the nixpkgs it crashed molecule, the last
released (non-master) version depends on pyyaml<6. Created a module to
implement multiple versions for pyyaml that is a candidate for
deprecation later; i.e. once molecule 3.6+ is released.